+++ ci_verify.sh
@@ -5,4 +5,5 @@ set -e
 
 pnpm -r build
 pnpm -r test
+pnpm -r lint
 pnpm exec prettier --check .