+++ webapp/src/plunder.d.ts
@@ -1,0 +1,9 @@
+import { type CaptureScreenshotManifest } from '@eighty4/plunder-core'
+
+declare global {
+    interface Window {
+        plunder: {
+            webpages: Array<CaptureScreenshotManifest>
+        }
+    }
+}