+++ cli/CHANGELOG.md
@@ -3,6 +3,7 @@
 - CLI updates progress of CSS parsing and screenshot capture
 - CLI command `devices` lists device labels
 - CLI command `links` checks anchor tags HREFs
+- CLI creates a web report to view captured screenshots
 
 ## 0.0.4 - 2025-02-10